Move calendar event handling to separate base class
